Fix for the B9 too many parts/plugins issue
These 2 small changes combined seem to fix it! 

Setting it to initialise instantly instead of on mainmenu..
Changing the ini_path to be relative to the executing assembly means the files can be placed anywhere.  

I've got mine in GameData\AltInput as that seems to be the standard.  (not sure if that reflection method is the most efficient way)
I think the order the plugins are discovered matters, it being in a directory beginning with 'A' puts it to the head of the queue.

It's only the dll location that matters but it seems odd to put the ini file somewhere else.